fix circular refs between TS files
