+++ src/parse.ts
@@ -3,7 +3,8 @@ import { AsyncAPIDocumentInterface, newAsyncAPIDocument } from './models';
 import { customOperations } from './custom-operations';
 import { validate } from "./lint";
 import { unfreeze } from './stringify';
-import { createDetailedAsyncAPI, normalizeInput, toAsyncAPIDocument } from "./utils";
+import { toAsyncAPIDocument } from "./document";
+import { createDetailedAsyncAPI, normalizeInput } from "./utils";
 
 import { xParserSpecParsed } from './constants';
